new component test added
