+++ src/Testing data/dropdowntestingdata.jsx
@@ -1,0 +1,14 @@
+const selects = [
+  {
+    name: "Settings",
+    url: "/settings",
+  },
+  {
+    name: "Logout",
+    onClick1() {
+      alert("Blah BLah");
+    },
+  },
+];
+
+export default selects;
